// eBibliotekaClient/src/app/data/interfaces/Librarian.ts
export interface RegisterLibrarianVM {
  email: string;
  username: string;
  password: string;
  firstName: string;
  lastName: string;
  libraryName: string;
}

export interface GetLibrarianVM {
  email: string;
  username: string;
  password: string;
  firstName: string;
  lastName: string;
}